Client - Refactor Code Smell
Rename device-detail-widget files to appropriate exports.
